move is_imported in init
